+++ deny.toml
@@ -10,7 +10,6 @@ allow = [
     "BSD-2-Clause",
     "BSD-3-Clause",
     "ISC",
-    "Unicode-DFS-2016",
     "MPL-2.0",
     "Unicode-3.0",
 ]
@@ -20,9 +19,5 @@ unknown-registry = "warn"
 unknown-git = "warn"
 
 [advisories]
-# Ignore unmaintained dev dependencies in criterion
-ignore = [
-    "RUSTSEC-2021-0145",  # atty unaligned read (dev dependency)
-    "RUSTSEC-2024-0375",  # atty unmaintained (dev dependency)
-    "RUSTSEC-2021-0127",  # serde_cbor unmaintained (dev dependency)
-]
+# No specific advisories to ignore currently
+ignore = []
